RENAME Kafka 및 Feign에 쓰이는 DTO 폴더 이동
