+++ src/main/java/com/jeontongju/authentication/security/oauth/OAuth2AuthenticationSuccessHandler.java
@@ -7,7 +7,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.jeontongju.authentication.dto.ResponseFormat;
+import com.jeontongju.authentication.dto.temp.ResponseFormat;
 import com.jeontongju.authentication.dto.response.JwtAccessTokenResponse;
 import com.jeontongju.authentication.security.MemberDetails;
 import com.jeontongju.authentication.security.jwt.JwtTokenProvider;
