+++ src/main/java/com/jeontongju/authentication/service/feign/consumer/ConsumerServiceClient.java
@@ -1,7 +1,7 @@
 package com.jeontongju.authentication.service.feign.consumer;
 
-import com.jeontongju.authentication.dto.request.ConsumerInfoForCreateBySnsRequestDto;
-import com.jeontongju.authentication.dto.request.ConsumerInfoForCreateRequestDto;
+import com.jeontongju.authentication.dto.temp.ConsumerInfoForCreateBySnsRequestDto;
+import com.jeontongju.authentication.dto.temp.ConsumerInfoForCreateRequestDto;
 import com.jeontongju.authentication.dto.temp.FeignFormat;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
